zipstream and streaming input optimizations (#5)
* Added initial output stream buffering

* zip/output streaming: checking if file exists before setting timestamp; minor cleanups

* Vastly optimized extracting time for small files - fixed slow console prints

* The progress bar is updated less frequently; Downloading of a file is aborted if error occurs or the page reloads; A root item in treeview has been added to simplify checking/unchecking all files.

* Selecting first file at list has been fixed - JS undefined condition.

* Removed libzip from deps, as it's no longer used

---------

+++ src/wasm/emjs.h
@@ -17,7 +17,7 @@ void down_wrap(std::string const &filename);
 void ui_innerhtml(const char *id, const char *value);
 void ui_setattr(const char *id, const char *attr, std::string const &value);
 void ui_remattr(const char *id, const char *attr);
-void ui_progbar_update(float value);
+void ui_progbar_update(int value);
 void ui_show_error();
 
 void open(const char *name, const char *__restrict modes);
